UserWishlist controller, service and repository

+++ ecommerce/src/main/java/com/ecommerce/repositories/implementations/HistoryRepositoryImplementation.java
@@ -1,5 +1,6 @@
 package com.ecommerce.repositories.implementations;
 
+import com.ecommerce.exceptions.NotFoundException;
 import com.ecommerce.models.CartItems;
 import com.ecommerce.models.CompositePK;
 import com.ecommerce.models.HistoryItems;
@@ -49,7 +50,20 @@ public class HistoryRepositoryImplementation implements HistoryRepository {
     public void deleteHistoryItem(Long userId, Long productId) {
         Session session = sessionFactory.getCurrentSession();
         HistoryItems historyItem = session.get(HistoryItems.class, new CompositePK(userId, productId));
+        if (historyItem == null) {
+            throw new NotFoundException();
+        }
         session.delete(historyItem);
+    }
+
+    @Override
+    public HistoryItems findById(Long userId, Long productId) {
+        Session session = sessionFactory.getCurrentSession();
+        HistoryItems historyItem = session.get(HistoryItems.class, new CompositePK(userId, productId));
+        if (historyItem == null) {
+            throw new NotFoundException();
+        }
+        return historyItem;
     }
 
     private HistoryItems convertFromCartToHistory(CartItems cartItem) {
@@ -60,6 +74,4 @@ public class HistoryRepositoryImplementation implements HistoryRepository {
         historyItem.setPurchaseDate(LocalDate.now());
         return historyItem;
     }
-
-
 }
